Modifying logic for PD tests to run

+++ tests/pingdirectory/01-ldap-tests.sh
@@ -9,9 +9,8 @@ if skipTest "${0}"; then
 fi
 
 testLdap() {
-  log "IS_BELUGA_ENV set to: ${IS_BELUGA_ENV}"
   log "Using the hostname: ${PINGDIRECTORY_ADMIN} and port: ${PD_SEED_LDAPS_PORT}"
-  if test "${IS_BELUGA_ENV}" = 'true'; then
+  if test "${IS_BELUGA_ENV}" = 'true' && test "${CI_SERVER}" != "yes"; then
     log "Running process grep and ldapsearch..."
     if pgrep -f docker > /dev/null; then
       docker run --rm pingidentity/ldap-sdk-tools ldapsearch \
